fix: stock reservation not working for sales invoice with update stock
